added svg
the drawing as a sketch -not finished yet.

Still need to figure out how to implement best in the hugo style
need to add some diagrams for the different scales
or at least a legend of some sorts that explains which
length scale is typical for particle, nuclear, atom physics etc.